dev: use latest typst-dom
